js setUTCMonth is stupid
Date.prototype.setUTCMonth()
[...]
Parameters:
monthValue
An integer between 0 and 11, representing the months January through December.
dayValue
Optional. An integer from 1 to 31, representing the day of the month.

problem only occurs if the current day of the month is a day that
doesn't exist for the previous month
problem now fixed