improve infoBox positioning
Prior calculation of the top-left corner location of the infoBox could
result in negative values, placing the element (partially) out of view.
Also, positioning was inconsistent, depending on whether left/right or
above/below the plane marker.

With this change, the size of the infoBox is taken info account when
calculating the relative position, and it ensured that the values would
not be less than zero.