allow multi range outlines to be used
This is a special case for a tar1090 instance that shows the combined output of
multiple readsb instances (like in a stage2 instance of the adsb.im feeder
image).

This tries once to download a multiOutline.json file and remembers if it exists
or if it should use the outline.json file instead. In the update loop it simply
checks for a third format where points is an array of arrays. And then loops
over that outer array - which in the case of the regular output.json is simply
an array with just one entry.

The additional overhead is negligible, the additional feature is really cool
(once you have the code to create merged outlines).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>