Swap Shipname & Callsign
Current tar1090 displays the ship's AIS callsign on the map. On aircraft, the callsign displays the flight number or registration on the map and is usable. 

For ships, displaying the ship name on the map is more friendly and usable. This code proposes swapping pr.shipname to ac.flight and pr.callsign to ac.r