allow host-relative link for aiscatcher_server
This is analog to what we do for the configLink - simply support the magic
token HOSTNAME to represent the request host.

That way a reverse proxy on the system running tar1090 can allow access to
geojson and icons.png regardless of where the ais_cather is running, massively
simplifying configuration.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>